Adds a new 'speculative transactions required to fix' feature
This is because certain people @p--b are absolutely awful at actually
bothering to accept purchases.

It basically assumes that if a purchase is incomplete, then the
debt should be spread equally between the people who haven't yet
accepted it. This isn't the /best/ heuristic in the world, but in
most cases it works just fine.